refactor: Restore migration and add shared timeline fetch logic
This commit addresses feedback from a code review on the initial UI refactoring.

It restores the data migration functionality that was lost and prepares for the removal of duplicated code by creating a shared data-fetching function.

- Re-implements the `migrate_data_from_files` function in `main.rs` to handle migrating data from the old file-based cache to the new LMDB database.
- Adds a new shared `fetch_timeline_events` async function to `nostr_client.rs` to encapsulate the duplicated logic for fetching timeline posts.

Note: The view modules have not yet been updated to use this new shared function. This is a partial implementation of the code duplication fix.
