+++ src/main.rs
@@ -1,241 +1,87 @@
 mod cache_db;
 mod nostr_client;
 mod ui;
+mod types;
 
 use eframe::egui;
-use nostr::{Keys, PublicKey};
-use nostr_sdk::Client;
-
 use std::collections::{HashMap, HashSet};
 use std::path::Path;
 use std::sync::{Arc, Mutex};
 use tokio::runtime::Runtime;
+use std::fs;
+use nostr::PublicKey;
+use regex::Regex;
 
-use self::nostr_client::{
-    connect_to_relays_with_nip65, fetch_nip01_profile, fetch_relays_for_followed_users,
-};
-use crate::cache_db::LmdbCache;
-use crate::ui::migrate_data_from_files;
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
+mod theme;
+
+use crate::cache_db::{LmdbCache, DB_FOLLOWED, DB_PROFILES, DB_RELAYS};
+use crate::types::*;
+
 
 const CONFIG_FILE: &str = "config.json"; // 設定ファイル名
-const CACHE_DIR: &str = "cache"; // Kept for migration
+
 const DB_PATH: &str = "cache_db";
-const CACHE_TTL_SECONDS: i64 = 24 * 60 * 60; // 24 hours
+const CACHE_DIR: &str = "cache"; // Re-added for migration
 
 const MAX_STATUS_LENGTH: usize = 140; // ステータス最大文字数
 
-#[derive(Serialize, Deserialize)]
-pub struct Config {
-    encrypted_secret_key: String, // NIP-49フォーマットの暗号化された秘密鍵
-    salt: String,                 // PBKDF2に使用するソルト (Base64エンコード)
-}
+async fn migrate_data_from_files(
+    cache_db: &LmdbCache,
+) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    let cache_path = Path::new(CACHE_DIR);
+    if !cache_path.exists() {
+        return Ok(());
+    }
 
-// --- キャッシュデータ構造 ---
-#[derive(Serialize, Deserialize, Clone)]
-pub struct Cache<T> {
-    pub timestamp: DateTime<Utc>,
-    pub data: T,
-}
+    println!("Old cache directory found. Starting data migration...");
 
-impl<T> Cache<T> {
-    pub fn new(data: T) -> Self {
-        Self {
-            timestamp: Utc::now(),
-            data,
+    let mut files_by_pubkey: HashMap<String, Vec<std::path::PathBuf>> = HashMap::new();
+    let re = Regex::new(r"([a-f0-9]{64})_.*\.json")?;
+
+    for entry in fs::read_dir(cache_path)? {
+        let entry = entry?;
+        let path = entry.path();
+        if path.is_file() {
+            if let Some(captures) = re.captures(path.file_name().unwrap().to_str().unwrap()) {
+                if let Some(pubkey) = captures.get(1) {
+                    files_by_pubkey
+                        .entry(pubkey.as_str().to_string())
+                        .or_default()
+                        .push(path);
+                }
+            }
         }
     }
 
-    pub fn is_expired(&self) -> bool {
-        let now = Utc::now();
-        let duration = now.signed_duration_since(self.timestamp);
-        duration.num_seconds() > CACHE_TTL_SECONDS
-    }
-}
-
-// NIP-01 プロファイルメタデータのための構造体
-// フィールドはNIP-01の推奨に従う
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct ProfileMetadata {
-    #[serde(default)]
-    pub name: String,
-    #[serde(default)]
-    pub about: String,
-    #[serde(default)]
-    pub picture: String,
-    #[serde(default)]
-    pub nip05: String, // NIP-05 identifier
-    #[serde(default)]
-    pub lud16: String, // Lightning Address
-    #[serde(flatten)] // その他の不明なフィールドを保持
-    pub extra: std::collections::HashMap<String, serde_json::Value>,
-}
-
-// リレーリスト編集のための構造体
-#[derive(Debug, Clone, Default)]
-pub struct EditableRelay {
-    pub url: String,
-    pub read: bool,
-    pub write: bool,
-}
-
-// 画像の読み込み状態を管理するenum
-#[derive(Clone)]
-pub enum ImageState {
-    Loading,
-    Loaded(egui::TextureHandle),
-    Failed,
-}
-
-// タイムラインの各投稿を表すための構造体
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TimelinePost {
-    pub author_pubkey: PublicKey,
-    pub author_metadata: ProfileMetadata,
-    pub content: String,
-    pub created_at: nostr::Timestamp,
-    #[serde(default)]
-    pub emojis: HashMap<String, String>, // shortcode -> url
-}
-
-// アプリケーションの内部状態を保持する構造体
-pub struct NostrStatusAppInternal {
-    pub cache_db: LmdbCache,
-    pub is_logged_in: bool,
-    pub status_message_input: String, // ユーザーが入力するステータス
-    pub show_post_dialog: bool,       // 投稿ダイアログの表示状態
-    pub secret_key_input: String,     // 初回起動時の秘密鍵入力用
-    pub passphrase_input: String,
-    pub confirm_passphrase_input: String,
-    pub nostr_client: Option<Client>,
-    pub my_keys: Option<Keys>,
-    pub followed_pubkeys: HashSet<PublicKey>, // NIP-02で取得したフォローリスト
-    pub followed_pubkeys_display: String,     // GUI表示用の文字列
-    pub timeline_posts: Vec<TimelinePost>,    // GUI表示用のステータスタイムライン
-    pub should_repaint: bool,                 // UIの再描画をトリガーするためのフラグ
-    pub is_loading: bool,                     // 処理中であることを示すフラグ
-    pub current_tab: AppTab,                  // 現在選択されているタブ
-    pub connected_relays_display: String,     // 接続中のリレー表示用
-    pub nip01_profile_display: String,        // GUI表示用のNIP-01プロファイルJSON文字列
-    pub editable_profile: ProfileMetadata,    // 編集可能なNIP-01プロファイルデータ
-    pub profile_fetch_status: String,         // プロファイル取得状態メッセージ
-    // リレーリスト編集用のフィールド
-    pub nip65_relays: Vec<EditableRelay>,
-    pub discover_relays_editor: String,
-    pub default_relays_editor: String,
-    pub current_theme: AppTheme,
-    pub image_cache: HashMap<String, ImageState>,
-}
-
-// タブの状態を管理するenum
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-pub enum AppTab {
-    Home, // ログイン/登録画面とタイムラインを含む
-    Relays,
-    Profile,
-}
-
-// UIテーマを管理するenum
-#[derive(Debug, PartialEq, Eq, Copy, Clone, Serialize, Deserialize)]
-pub enum AppTheme {
-    Light,
-    Dark,
-}
-
-impl AppTheme {
-    pub fn card_background_color(&self) -> egui::Color32 {
-        match self {
-            AppTheme::Light => egui::Color32::from_white_alpha(250),
-            AppTheme::Dark => egui::Color32::from_rgb(44, 44, 46),
+    for (pubkey_hex, paths) in files_by_pubkey {
+        println!("Migrating data for pubkey: {}", pubkey_hex);
+        for path in paths {
+            let file_name = path.file_name().unwrap().to_str().unwrap();
+            if file_name.ends_with("_followed.json") {
+                let content = fs::read_to_string(&path)?;
+                let cache: Cache<HashSet<PublicKey>> = serde_json::from_str(&content)?;
+                cache_db.write_cache(DB_FOLLOWED, &pubkey_hex, &cache.data)?;
+                println!("  - Migrated followed list.");
+            } else if file_name.ends_with("_nip65.json") {
+                let content = fs::read_to_string(&path)?;
+                let cache: Cache<Vec<(String, Option<String>)>> = serde_json::from_str(&content)?;
+                cache_db.write_cache(DB_RELAYS, &pubkey_hex, &cache.data)?;
+                println!("  - Migrated NIP-65 relays.");
+            } else if file_name.ends_with("_profile.json") {
+                let content = fs::read_to_string(&path)?;
+                let cache: Cache<ProfileMetadata> = serde_json::from_str(&content)?;
+                cache_db.write_cache(DB_PROFILES, &pubkey_hex, &cache.data)?;
+                println!("  - Migrated profile metadata.");
+            }
         }
     }
 
-    pub fn text_color(&self) -> egui::Color32 {
-        match self {
-            AppTheme::Light => egui::Color32::BLACK,
-            AppTheme::Dark => egui::Color32::WHITE,
-        }
-    }
-}
+    // Rename the old cache directory to prevent re-migration
+    let migrated_path = Path::new("cache_migrated");
+    fs::rename(cache_path, migrated_path)?;
+    println!("Data migration complete. Old cache directory renamed to 'cache_migrated'.");
 
-// --- ライトモードのVisualsを返す関数 ---
-pub fn light_visuals() -> egui::Visuals {
-    let mut visuals = egui::Visuals::light();
-    let background_color = egui::Color32::from_rgb(242, 242, 247);
-    let panel_color = egui::Color32::from_rgb(255, 255, 255);
-    let text_color = egui::Color32::BLACK;
-    let accent_color = egui::Color32::from_rgb(0, 110, 230);
-    let separator_color = egui::Color32::from_gray(225);
-
-    visuals.window_fill = background_color;
-    visuals.panel_fill = panel_color;
-    visuals.override_text_color = Some(text_color);
-    visuals.hyperlink_color = accent_color;
-    visuals.faint_bg_color = background_color;
-    visuals.extreme_bg_color = egui::Color32::from_gray(230);
-    visuals.window_stroke = egui::Stroke::new(1.0, separator_color);
-    visuals.selection.bg_fill = accent_color.linear_multiply(0.3);
-    visuals.selection.stroke = egui::Stroke::new(1.0, text_color);
-
-    let widget_visuals = &mut visuals.widgets;
-    widget_visuals.noninteractive.bg_fill = egui::Color32::TRANSPARENT;
-    widget_visuals.noninteractive.bg_stroke = egui::Stroke::NONE;
-    widget_visuals.noninteractive.fg_stroke = egui::Stroke::new(1.0, text_color);
-
-    widget_visuals.inactive.bg_fill = egui::Color32::from_gray(235);
-    widget_visuals.inactive.bg_stroke = egui::Stroke::NONE;
-    widget_visuals.inactive.fg_stroke = egui::Stroke::new(1.0, text_color);
-
-    widget_visuals.hovered.bg_fill = egui::Color32::from_gray(220);
-    widget_visuals.hovered.bg_stroke = egui::Stroke::NONE;
-    widget_visuals.hovered.fg_stroke = egui::Stroke::new(1.0, text_color);
-
-    widget_visuals.active.bg_fill = egui::Color32::from_gray(210);
-    widget_visuals.active.bg_stroke = egui::Stroke::NONE;
-    widget_visuals.active.fg_stroke = egui::Stroke::new(1.0, accent_color);
-
-    visuals
-}
-
-// --- ダークモードのVisualsを返す関数 ---
-pub fn dark_visuals() -> egui::Visuals {
-    let mut visuals = egui::Visuals::dark();
-    let background_color = egui::Color32::from_rgb(29, 29, 31);
-    let panel_color = egui::Color32::from_rgb(44, 44, 46);
-    let text_color = egui::Color32::from_gray(230);
-    let accent_color = egui::Color32::from_rgb(10, 132, 255);
-    let separator_color = egui::Color32::from_gray(58);
-
-    visuals.window_fill = background_color;
-    visuals.panel_fill = panel_color;
-    visuals.override_text_color = Some(text_color);
-    visuals.hyperlink_color = accent_color;
-    visuals.faint_bg_color = background_color;
-    visuals.extreme_bg_color = egui::Color32::from_gray(60);
-    visuals.window_stroke = egui::Stroke::new(1.0, separator_color);
-    visuals.selection.bg_fill = accent_color.linear_multiply(0.3);
-    visuals.selection.stroke = egui::Stroke::new(1.0, text_color);
-
-    let widget_visuals = &mut visuals.widgets;
-    widget_visuals.noninteractive.bg_fill = egui::Color32::from_rgb(40, 40, 40);
-    widget_visuals.noninteractive.bg_stroke = egui::Stroke::NONE;
-    widget_visuals.noninteractive.fg_stroke = egui::Stroke::new(1.0, text_color);
-
-    widget_visuals.inactive.bg_fill = egui::Color32::from_gray(50);
-    widget_visuals.inactive.bg_stroke = egui::Stroke::NONE;
-    widget_visuals.inactive.fg_stroke =
-        egui::Stroke::new(1.0, egui::Color32::from_rgb(180, 180, 180));
-
-    widget_visuals.hovered.bg_fill = egui::Color32::from_gray(70);
-    widget_visuals.hovered.bg_stroke = egui::Stroke::NONE;
-    widget_visuals.hovered.fg_stroke = egui::Stroke::new(1.0, text_color);
-
-    widget_visuals.active.bg_fill = egui::Color32::from_gray(85);
-    widget_visuals.active.bg_stroke = egui::Stroke::NONE;
-    widget_visuals.active.fg_stroke = egui::Stroke::new(1.0, accent_color);
-
-    visuals
+    Ok(())
 }
 
 // eframe::Appトレイトを実装する構造体
@@ -276,7 +122,7 @@ impl NostrStatusApp {
         _cc.egui_ctx.set_fonts(fonts);
 
         // --- スタイル調整 ---
-        style.visuals = light_visuals(); // ライトモードを基準にする
+        style.visuals = theme::light_visuals(); // ライトモードを基準にする
 
         // 角丸やテキストスタイルは共通で設定
         let corner_radius = 6.0;
